mass_edits/import_adries_minv/requirements.txt
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
click-plugins==1.1.1
cligj==0.7.2
conflate==0.5.2
Fiona==1.8.21
geojson==2.5.0
geopandas==0.10.2
idna==3.3
Jinja2==3.1.1
kdtree==0.16
lxml==4.8.0
MarkupSafe==2.1.1
munch==2.5.0
numpy==1.22.3
osm-conflate==1.4.1
overpass==0.7
pandas==1.4.2
pyproj==3.3.0
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.27.1
Rtree==1.0.0
Shapely==1.8.1.post1
six==1.16.0
urllib3==1.26.9
